♻️ Refactor: 유저 데이터 정보를 HeadNav의 프로필 컴포넌트에서 처리하도록 함
