+++ src/components/layout/HeadNav/Profile/Profile.tsx
@@ -1,11 +1,15 @@
 import { useEffect, useState } from "react";
-import { useCurrentUser } from "@/src/context/UserContext";
+import { useCurrentUser, useSetCurrentUser } from "@/src/context/UserContext";
 import Link from "next/link";
 import Image from "next/image";
 import * as S from "./Profile.style";
+import { acceptDataFromApi } from "@/src/utils/api";
+
+const USER = "users/1";
 
 export default function HeadNavProfile() {
   const userData = useCurrentUser();
+  const setCurrentUser = useSetCurrentUser();
 
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [accountEmail, setAccountEmail] = useState("");
@@ -13,15 +17,29 @@ export default function HeadNavProfile() {
     "/assets/icons/svg/nav-profile-default.svg"
   );
 
+  const accountVerification = async (user: string) => {
+    const receivedData = await acceptDataFromApi(user);
+    if (!receivedData) return;
+
+    const { data } = receivedData;
+    setCurrentUser(...data);
+  };
+
   useEffect(() => {
-    const hasAccessToken = Boolean(localStorage.getItem("accessToken"));
-    if (hasAccessToken && userData) {
+    if (userData) {
       setIsLoggedIn(true);
       setAccountEmail(userData.email);
       setProfileImg(userData.image_source);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [userData]);
+
+  useEffect(() => {
+    const hasAccessToken = Boolean(localStorage.getItem("accessToken"));
+    if (hasAccessToken) {
+      accountVerification(USER);
+    }
+  }, []);
 
   return (
     <>
